Fixes cabal file to work with stack
`stack` could not resolve the needed build artifacts without adding
`elm-package` as a dependency to the corresponding executable.
Additionally, adds Paths_elm_compiler to other-modules in order to
silence a warning from stack, and removes the `-threaded` build
option from the library to silence a warning from `cabal check`
about it doing nothing for libraries.