Move towards structured errors.
This will set things up for nicer formatting (and JSON output in the
long run).

Also, no longer require git to be present. It looks like we do not call
that anymore. I think it was no longer needed when we started
downloading .zip files directly.

This also stops using FlexibleContexts in as many places as possible.
Just being concrete.

It also uses a single HTTP manager for all the requests. This should
make repeated requests to github or package.elm-lang.org faster. Not
sure by how much though!